Changed int type for the zero point Small documentation fix

+++ src/ZMPT101B.cpp
@@ -44,7 +44,7 @@ float ZMPT101B::calibrateVoltage() {
 	return Vzero;
 }
 
-void ZMPT101B::setZeroPoint(int _zero) {
+void ZMPT101B::setZeroPoint(uint16_t _zero) {
 	zero = _zero;
 }
